requirements.txt
rembg
opencv-python
numpy
matplotlib
Pillow
